refactor(ci): share the strict artifact ZIP extractor
